Store checkpoint samples in float64
Our constraint residuals are 2e-16 on S^2 and 3.8e-14 on St(4,2). float32
storage has an epsilon of 1.2e-07, so checkpointing samples at float32
destroyed the very result the checkpoint exists to evidence, and left behind a
number that coincidentally resembles R-ASBS's 3.4e-07 retraction error.

common.py:save_ckpt now promotes any floating-point sample tensor to float64
and leaves integer state indices alone; the six call sites that pre-cast to
float32 no longer do. Network weights stay float32, which is what was trained.

gallery.py figure 7 gains the orthogonality panel back, but draws it only when
the checkpoint on disk is genuinely float64 -- an old float32 file degrades to
a printed notice rather than to a plausible-looking wrong panel.

regen_f64.sh refreshes the beta = 2 pair at the same config and seed.

+++ structured_asbs/sphere.py
@@ -492,13 +492,13 @@ def run_exact(args):
         out[steps] = m
         report(f"steps={steps:4d}", m)
         C.save_ckpt(args.ckpt_dir, f"{args.tag}_exact_steps{steps}",
-                    samples=x.to(torch.float32), extra={"metrics": m})
+                    samples=x, extra={"metrics": m})
     prob = SphereProblem(sigma=args.sigma, tau=args.tau, steps=128)
     torch.manual_seed(args.seed)
     xr = simulate(prob, None, args.n_samples, 128)
     report("reference   ", z_metrics(xr))
     C.save_ckpt(args.ckpt_dir, f"{args.tag}_reference",
-                samples=xr.to(torch.float32),
+                samples=xr,
                 extra={"metrics": z_metrics(xr)})
     # finite-sample KS floor
     grid, cdf = C.exact_s2_z_cdf_grid(num=40001)
@@ -639,7 +639,7 @@ def run_train(args):
                 xs_fin = simulate(prob, _score_fn(ema_net, steps),
                                   args.n_samples, steps)
             p = C.save_ckpt(args.ckpt_dir, f"{args.tag}_seed{seed}",
-                            net=ema_net, samples=xs_fin.to(torch.float32),
+                            net=ema_net, samples=xs_fin,
                             extra={"config": vars(args), "metrics": m,
                                    "score_rel_err": errs})
             print(f"    ckpt -> {p}", flush=True)
